# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SAHOOL Platform API",
    "description": "Agricultural Intelligence Platform - منصة الذكاء الزراعي",
    "version": "15.3.0",
    "contact": {
      "name": "SAHOOL Support",
      "email": "support@sahool.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8095",
      "description": "SAHOOL Agro Advisor - Disease diagnosis, nutrient assessment, and fertilizer planning for Yemen agriculture"
    },
    {
      "url": "http://localhost:8000",
      "description": "SAHOOL Agent Registry Service - Agent card to register"
    },
    {
      "url": "http://localhost:8095",
      "description": "SAHOOL Agro Advisor - Disease diagnosis, nutrient assessment, and fertilizer planning for Yemen agriculture"
    },
    {
      "url": "http://localhost:8000",
      "description": "AI Advisor Service - User question"
    },
    {
      "url": "http://localhost:8113",
      "description": "SAHOOL Alert Service - معرف التنبيه"
    },
    {
      "url": "http://localhost:8111",
      "description": "SAHOOL Astronomical Calendar Service - مفتاح المرحلة"
    },
    {
      "url": "http://localhost:8089",
      "description": "SAHOOL Billing Core | خدمة الفوترة - Complete billing, subscription, and payment management for SAHOOL platform"
    },
    {
      "url": "http://localhost:8100",
      "description": "SAHOOL Crop Health Service - Normalized Difference Vegetation Index"
    },
    {
      "url": "http://localhost:8095",
...
  "paths": {
...
    "/stats": {
      "get": {
        "summary": "Get gateway statistics including room and NATS information",
        "description": "",
        "tags": ["ws-gateway"],
        "x-summary-ar": "",
        "x-description-ar": "",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "websocket",
            "in": "path",
            "required": true,
            "description": "",
            "schema": {
              "type": "websocket"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/v1/stats": {
      "get": {
        "summary": "إحصائيات الإشعارات",
        "description": "",
        "tags": ["notification-service"],
        "x-summary-ar": "إحصائيات الإشعارات",
        "x-description-ar": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service communication"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}